requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
astor==0.8.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
filelock==3.12.0
fonttools==4.39.3
frozenlist==1.3.3
idna==3.4
importlib-resources==5.12.0
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lit==16.0.2
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
ply==3.11
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.16.2
pydot==1.4.2
pyfiglet==0.8.post1
Pygments==2.15.1
PyGObject==3.38.0
pygraphviz==1.10
pygwalker==0.1.9
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.29.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
threadpoolctl==3.1.0
torch==2.0.0
torch-geometric==2.3.1
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing-extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
xdot==1.2
yamlordereddictloader==0.4.0
yarl==1.9.2
zipp==3.15.0
